put transformer into separate file
