+++ tests/utils/test_tf_model_data.py
@@ -4,7 +4,7 @@ import pytest
 import scipy.sparse
 import numpy as np
 
-from rasa.utils.tensorflow.tf_model_data import RasaModelData
+from rasa.utils.tensorflow.model_data import RasaModelData
 
 
 @pytest.fixture
